requirement.txt
streamlit
numpy
streamlit-cropper
Pillow